♻️  Reorganizing compositions for better reusability
