+++ appcues/src/main/java/com/appcues/analytics/ExperienceLifecycleEvent.kt
@@ -3,7 +3,7 @@ package com.appcues.analytics
 import com.appcues.analytics.ExperienceLifecycleEvent.StepInteraction.InteractionType.FORM_SUBMITTED
 import com.appcues.data.model.Experience
 import com.appcues.statemachine.Error
-import com.appcues.ui.ExperienceStepFormState
+import com.appcues.ui.utils.ExperienceStepFormState
 import com.appcues.util.toSlug
 
 internal sealed class ExperienceLifecycleEvent(
