+++ appcues/src/main/java/com/appcues/trait/appcues/PagingDotsTrait.kt
@@ -26,9 +26,9 @@ import com.appcues.data.model.AppcuesConfigMap
 import com.appcues.data.model.getConfigStyle
 import com.appcues.trait.ContainerDecoratingTrait
 import com.appcues.trait.ContainerDecoratingTrait.ContainerDecoratingType
+import com.appcues.ui.composables.rememberAppcuesPaginationState
 import com.appcues.ui.extensions.getBoxAlignment
 import com.appcues.ui.extensions.getColor
-import com.appcues.ui.rememberAppcuesPaginationState
 
 internal class PagingDotsTrait(
     override val config: AppcuesConfigMap
